+++ src/main/java/com/thesarlaacsweep/radioactiveblocks/items/RadioactiveArmorItem.java
@@ -1,5 +1,6 @@
 package com.thesarlaacsweep.radioactiveblocks.items;
 
+import com.thesarlaacsweep.radioactiveblocks.config.ModConfig;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.inventory.EquipmentSlotType;
 import net.minecraft.item.ArmorItem;
@@ -16,13 +17,15 @@ public class RadioactiveArmorItem extends ArmorItem {
 
     @Override
     public void onArmorTick(ItemStack stack, World world, PlayerEntity player) {
-        if (
-                !player.getItemStackFromSlot(EquipmentSlotType.HEAD).isEmpty() &&
-                        !player.getItemStackFromSlot(EquipmentSlotType.CHEST).isEmpty() &&
-                        !player.getItemStackFromSlot(EquipmentSlotType.LEGS).isEmpty() &&
-                        !player.getItemStackFromSlot(EquipmentSlotType.FEET).isEmpty()
-        ) {
-            player.addPotionEffect(new EffectInstance(Effects.REGENERATION));
+        if (ModConfig.COMMON_CONFIG.has_radiation_effect.get()) {
+            if (
+                    !player.getItemStackFromSlot(EquipmentSlotType.HEAD).isEmpty() &&
+                            !player.getItemStackFromSlot(EquipmentSlotType.CHEST).isEmpty() &&
+                            !player.getItemStackFromSlot(EquipmentSlotType.LEGS).isEmpty() &&
+                            !player.getItemStackFromSlot(EquipmentSlotType.FEET).isEmpty()
+            ) {
+                player.addPotionEffect(new EffectInstance(Effects.REGENERATION));
+            }
         }
     }
 }